+++ src/components/description/index.js
@@ -1,0 +1,2 @@
+export { default as DescriptionSubscriber } from './DescriptionSubscriber';
+export { default as Description } from './Description';